fix: jwt long casting & update user status

+++ morning-holic/morning-holic-app/src/main/kotlin/com/morningholic/morningholicapp/securities/JwtUtils.kt
@@ -68,7 +68,7 @@ object JwtUtils {
 
     fun parseUserIdFromToken(token: String): Long {
         val claims = getAllClaims(token)
-        return claims["userId"] as Long
+        return (claims["userId"] as Int).toLong()
     }
 
     fun getAuthenticationByUserId(userId: Long): Authentication {
